COMP 584: Backend Week 8
TODO: Before week 9: fix my defect which is preventing countries being output to SwaggerUI webpage

+++ MyWebAPI/Data/WorldCitiesCSV.cs
@@ -4,8 +4,8 @@
     {
         public string city { get; set; } = null!;
         public string city_ascii { get; set; } = null!;
-        public decimal lat { get; set; }
-        public decimal lng { get; set; }
+        public double lat { get; set; }
+        public double lng { get; set; }
         public string country { get; set; } = null!;
         public string iso2 { get; set; } = null!;
         public string iso3 { get; set; } = null!;
